Подправил текущую реализацию
