+++ cs/Markdown/InlineParser.cs
@@ -1,0 +1,77 @@
+﻿using Markdown.Interfaces;
+using System.Text;
+
+namespace Markdown;
+
+public class InlineParser
+{
+    public IReadOnlyList<INode> Parse(string text)
+    {
+        // TODO: реализовать метод
+        throw new NotImplementedException();
+    }
+
+    /// <summary>
+    /// Заменяет экранированные символы (\_, \#, \\) на временные маркеры.
+    /// </summary>
+    private static string PreprocessEscapes(string text)
+    {
+        return string.Empty;
+    }
+
+    /// <summary>
+    /// Длина последовательности символов подчеркивания
+    /// </summary>
+    private static int GetUnderlineLength(string text, int position)
+    {
+        return 0;
+    }
+
+    /// <summary>
+    /// Может ли текущий символ подчеркивания открыть выделение
+    /// </summary>
+    private static bool CanOpen(string text, int position, int length)
+    {
+        return false;
+    }
+
+    /// <summary>
+    /// Может ли текущий символ подчеркивания закрыть выделение
+    /// </summary>
+    private static bool CanClose(string text, int position, int length)
+    {
+        return false;
+    }
+
+    /// <summary>
+    /// Добавляет текст из буфера в список узлов, если буфер не пуст
+    /// </summary>
+    private static void AddTextIfBufferNotEmpty(List<INode> nodes, StringBuilder textBuffer)
+    {
+        throw new NotImplementedException();
+    }
+
+    /// <summary>
+    /// Добавляет текстовый узел в список узлов
+    /// </summary>
+    private static void AddTextNode(List<INode> nodes, string text)
+    {
+        throw new NotImplementedException();
+    }
+
+    /// <summary>
+    /// Возвращает временный маркер в исходный символ
+    /// </summary>
+    private static void RestorePlaceholders(List<INode> nodes)
+    {
+        throw new NotImplementedException();
+    }
+
+    /// <summary>
+    /// Объединяет соседние текстовые узлы в один
+    /// </summary>
+    private static void CombineNeighborTextNodes(List<INode> nodes)
+    {
+        throw new NotImplementedException();
+    }
+}
