+++ cs/Markdown/Interfaces/IBlock.cs
@@ -1,0 +1,16 @@
+﻿namespace Markdown.Interfaces;
+
+public interface IBlock
+{
+    string RawText { get; }
+
+    BlockType Type { get; }
+
+    public IReadOnlyList<INode> Inlines { get; set; }
+}
+
+public enum BlockType
+{
+    Heading,
+    Paragraph
+}
