+++ cs/Markdown/Md.cs
@@ -1,9 +1,19 @@
 ﻿namespace Markdown;
 
-public class Md(Pipeline pipeline)
+public class Md
 {
-    public string Render(string markdownText)
+    public string Render(string text, BlockSegmenter segmenter, InlineParser parser, HtmlRenderer renderer)
     {
-        return pipeline.Run(markdownText);
+        var blocks = segmenter.Segment(text);
+
+        foreach (var block in blocks)
+        {
+            var inlines = parser.Parse(block.RawText);
+            block.Inlines = inlines;
+        }
+
+        var html = renderer.Render(blocks);
+
+        return html;
     }
 }
