--- cs/Markdown/Parsing/InlineParser.cs
@@ -1,19 +1,0 @@
-﻿using Markdown.Nodes;
-
-namespace Markdown.Parsing;
-
-public class InlineParser
-{
-    public IReadOnlyList<INode> Parse(string text)
-    {
-        // TODO: реализовать метод
-        throw new NotImplementedException();
-    }
-
-    // Предполагается реализация с разделением кода парсинга на отдельные классы, работающие в следующем порядке:
-    // Класс, посимвольно читающий текст
-    // Класс, преобразующий текст в единицы разметки (текст, бэк слеш, одинарное подчеркивание итд)
-    // Класс, отвечающий за обработку экранирования в тексте после преобразования текста в единицы разметки
-    // Класс, проверяющий, можно ли считать подчеркивание началом или концом выделения фрагмента текста. 
-    // Класс, определяющий, что именно нужно выделить курсивом, жирным или оставить как есть. Формирует итоговый результат.
-}
